requirements.txt
# OCR
pytesseract
easyocr
opencv-python
Flask

# ML
torch
torchvision
scikit-learn
numpy
Pillow
albumentationsx
ultralytics

# NLP / PII
spacy
regex

# API (planned)
fastapi
uvicorn

# Utilities
python-dotenv
tqdm
matplotlib

# Testing
pytest
